# backend/docs/openapi.v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects/{projectId}": {
      "get": {
        "operationId": "ProjectsController_get",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ProjectsController_update",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/projects/{projectId}/documents/presign-upload": {
      "post": {
        "operationId": "DocumentsController_presign",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/documents/{id}/confirm": {
      "post": {
        "operationId": "DocumentsController_confirm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/projects/{projectId}/documents": {
      "get": {
        "operationId": "DocumentsController_list",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stageId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stepId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "contract",
                "act",
                "estimate",
                "warranty",
                "photo",
                "blueprint",
                "other"
              ],
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/documents/{id}": {
      "get": {
        "operationId": "DocumentsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "DocumentsController_patch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "DocumentsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/documents/{id}/download": {
      "get": {
        "operationId": "DocumentsController_download",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/documents/{id}/thumbnail": {
      "get": {
        "operationId": "DocumentsController_thumbnail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/notification-settings": {
      "get": {
        "operationId": "NotificationsController_getSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "NotificationsController_patchSetting",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchSettingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/notification-logs": {
      "get": {
        "operationId": "NotificationsController_logs",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/admin/projects": {
      "get": {
        "operationId": "AdminProjectsController_list",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "active",
                "archived"
              ],
              "type": "string"
            }
          },
          {
            "name": "ownerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/projects/{id}": {
      "get": {
        "operationId": "AdminProjectsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/projects/{id}/force-archive": {
      "post": {
        "operationId": "AdminProjectsController_forceArchive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForceArchiveDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/admin/broadcasts/preview": {
      "post": {
        "operationId": "BroadcastsController_preview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-broadcasts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/broadcasts": {
      "post": {
        "operationId": "BroadcastsController_send",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendBroadcastDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-broadcasts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "BroadcastsController_list",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-broadcasts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/broadcasts/{id}": {
      "get": {
        "operationId": "BroadcastsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-broadcasts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Repair Control API",
    "description": "OpenAPI v1.0 frozen contract. Источник истины для Flutter retrofit-клиентов.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "plannedStart": {
            "type": "string"
          },
          "plannedEnd": {
            "type": "string"
          },
          "workBudget": {
            "type": "integer"
          },
          "materialsBudget": {
            "type": "integer"
          }
        }
      },
...
      "PresignUploadDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "contract",
              "act",
              "estimate",
              "warranty",
              "photo",
              "blueprint",
              "other"
            ]
          },
          "title": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "number"
          },
          "stageId": {
            "type": "string"
          },
          "stepId": {
            "type": "string"
          }
        },
        "required": [
          "category",
          "title",
          "mimeType",
          "sizeBytes"
        ]
      },
      "PatchDocumentDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "contract",
              "act",
              "estimate",
              "warranty",
              "photo",
              "blueprint",
              "other"
            ]
          },
          "stageId": {
            "type": "string"
          },
          "stepId": {
            "type": "string"
          }
        }
      },
      "PatchSettingDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "approval_requested",
              "approval_approved",
              "approval_rejected",
              "payment_created",
              "payment_confirmed",
              "payment_disputed",
              "payment_resolved",
              "stage_rejected_by_customer",
              "stage_overdue",
              "stage_deadline_exceeds_project",
              "material_request_created",
              "material_delivered",
              "material_disputed",
              "selfpurchase_created",
              "tool_issued",
              "export_completed",
              "export_failed",
              "chat_message_new",
              "step_completed",
              "stage_completed",
              "stage_paused",
              "note_created_for_me",
              "question_asked",
              "project_archived",
              "membership_added"
            ]
          },
          "pushEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "kind",
          "pushEnabled"
        ]
      },
...
      "ForceArchiveDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "reason"
        ]
      },
...
      "BroadcastFilterDto": {
        "type": "object",
        "properties": {
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "customer",
                "representative",
                "contractor",
                "master",
                "admin"
              ]
            }
          },
          "projectIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "userIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bannedOnly": {
            "type": "boolean"
          }
        }
      },
      "PreviewDto": {
        "type": "object",
        "properties": {
          "filter": {
            "$ref": "#/components/schemas/BroadcastFilterDto"
          }
        },
        "required": [
          "filter"
        ]
      },
      "SendBroadcastDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "deepLink": {
            "type": "string"
          },
          "filter": {
            "$ref": "#/components/schemas/BroadcastFilterDto"
          }
        },
        "required": [
          "title",
          "body",
          "filter"
        ]
      }
    }
  }
}